### 2025-12-25 9a4dd92
cleanup code + fix missing dois
