+++ srunner/scenariomanager/scenario_manager.py
@@ -178,22 +178,18 @@ class ScenarioManager(object):
             if self._debug_mode:
                 print("\n--------- Tick ---------\n")
 
-            
+            if self._agent is not None:
+                self._agent()  # pylint: disable=not-callable
 
             _tick_carla_start = time.perf_counter_ns() / 1e6
             if self._sync_mode and self._watchdog.get_status():
                 CarlaDataProvider.get_world().tick()
+                GameTime.on_carla_tick(timestamp)
+                CarlaDataProvider.on_carla_tick()
 
             MetricsCollector.update_key(
                 "carla_time", (time.perf_counter_ns() / 1e6) - _tick_carla_start
             )
-            
-            if self._agent is not None:
-                self._agent()  # pylint: disable=not-callable
-
-            # Update game time and actor information
-            GameTime.on_carla_tick(timestamp)
-            CarlaDataProvider.on_carla_tick()
 
             # Tick scenario
             _scenario_tick_start = time.perf_counter_ns() / 1e6
