### 2025-12-26 e9b631f
moved executor to 2 threads; improves reliability
