### 2025-12-26 ea016a5
added new dds configs for local and dist
