### 2025-12-28 c8c1a00
updated DDS configs
