### 2025-12-29 bcee022
updated LICENSE, README
