+++ README.md
@@ -1,14 +1,13 @@
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
-![GitHub tag (latest SemVer)](https://img.shields.io/github/tag/carla-simulator/scenario_runner.svg)
 
 CAWSR: ScenarioRunner for CARLA with support for Autoware
 ========================
-This repository contains scenario definition and an execution engine
-for CARLA. Support has been added to run route-based scenarios with the ego being controlled by [Autoware](https://autoware.org/autoware-overview/)
+
+CAWSR (Carla Autoware Scenario Runner) is a scenario execution engine built for the testing of [Autoware](https://autoware.org/autoware-overview/) in route-based scenarios.
 
 Prerequisites
 ---------------------------
-Both CARLA and Autoware require a high-spec computer with a high-end Nvidia GPU. It is also possible to run a [**distributed**]() setup with multiple machines to help ease the workload. Currently, only Linux is supported (guide was written on Ubuntu 24.04).
+Both CARLA and Autoware require a high-spec computer with a high-end Nvidia GPU. It is also possible to run a [**distributed**]() setup with multiple machines to help ease the workload, or run the entire stack locally. Currently, only Linux is supported (guide was written on Ubuntu 24.04).
 
 Ensure the target machine(s) have the [Docker Engine]() and [Nvidia Container toolkit]() installed to enable gpu accelerated workflows in Docker.
 
@@ -23,7 +22,7 @@ docker pull ghcr.io/intelligent-testing-lab/autoware-scenario-runner:latest
 docker pull ghcr.io/intelligent-testing-lab/autoware:latest
 ```
 
-Autoware and ROS use a custom messaging interface for communication, known as DDS. They support various implementations, but they all rely on specific network settings to enable maximum data transfer. Save the following command in `setup.sh`, allow it to be executable `chmod +x setup.sh` and run.
+Autoware and ROS use a custom messaging interface for communication, known as DDS. For maximum performance, configure your network settings as follows. If not configured, you will see [heavy performance issues](https://docs.ros.org/en/humble/How-To-Guides/DDS-tuning.html#cross-vendor-tuning) as the default ubuntu buffer sizes fill up fast, especially when running over lossy networks such as WiFi.
 ```bash
 # Increase the maximum receive and send buffer size for network packets, allowing our containers to communicate
 sudo sysctl -w net.core.rmem_max=2147483647  # 2 GiB, default is 208 KiB
@@ -33,6 +32,7 @@ sudo sysctl -w net.core.wmem_max=2147483647
 sudo sysctl -w net.ipv4.ipfrag_time=3  # in seconds, default is 30 s
 sudo sysctl -w net.ipv4.ipfrag_high_thresh=134217728  # 128 MiB, default is 256 KiB
 ```
+Save the following commands in `setup.sh`, allow it to be executable `chmod +x setup.sh`.
 These settings are **temporary** and will revert on restart.
 
 To allow GUI applications (like Autoware and CARLA) to run through Docker, you must allow xhost connections from the `docker` group.
@@ -40,24 +40,95 @@ To allow GUI applications (like Autoware and CARLA) to run through Docker, you m
 xhost +local:docker
 ```
 
-Using CAWSR
+Running CAWSR
 ------------------------
 
-After completiting the prerequisite steps, clone the [CAWSR workspace](https://github.com/Intelligent-Testing-Lab/cawsr_workspace) repository. To launch CAWSR, navigate to the CAWSR workspace and run `docker compose up`.
+CAWSR can both be ran `locally` or `distributed`. Due to the high-spec requirements, it is recommended to run distributed if you do not meet the following minimum specs:
[…]
