### 2025-12-29 5a72683
cleaned up initialisation code
