### 2025-12-30 3a611b6
moved agent initialisation
