### 2025-12-30 dc853c9
removed GameTime Restart from load_scenario
