+++ cawsr.py
@@ -155,6 +155,9 @@ class AWScenarioRunner(object):
 
         logger.info(f"{settings.__str__()}")
 
+        logger.info("Restarting GameTime...")
+        GameTime.restart()
+
         # update the world
         CarlaDataProvider.set_world(self.carla_world)
 
@@ -193,6 +196,18 @@ class AWScenarioRunner(object):
 
         ego.prepare_ego(route[0][0])  # set location to first waypoint
 
+        # allow the agent X ticks to initialize sensors and set the route
+        logger.info("Initialising agent route...")
+        budget = self._conf["initialisation_budget"]
+        status = False
+        for tick in range(1, budget + 1):
+            status = self.aw_agent.run_step_init()  # type: ignore
+            self._tick_carla()
+        if not status:
+            logger.info("Agent failed to initialise route")
+        else:
+            logger.info("Successfully initialised agent; route set.")
+
         logger.info("Loading Traffic Manager...")
         tm_port = int(self._carla.TRAFFIC_MANAGER.PORT)  # type: ignore
         CarlaDataProvider.set_traffic_manager_port(tm_port)
@@ -212,18 +227,6 @@ class AWScenarioRunner(object):
         except Exception:
             logger.info("Could not load Route Scenario")
             traceback.print_exc()
-
-        # allow the agent X ticks to initialize sensors and set the route
-        logger.info("Initialising agent route...")
-        budget = self._conf["initialisation_budget"]
-        status = False
-        for tick in range(1, budget + 1):
-            status = self.aw_agent.run_step_init()  # type: ignore
-            self._tick_carla()
-        if not status:
-            logger.info("Agent failed to initialise route")
-        else:
-            logger.info("Successfully initialised agent; route set.")
 
         logger.info("Starting scenario...")
 
