### 2025-12-30 3324679
fixed bug where bridge wouldn't tick after route set
