### 2026-01-06 e64f003
Fix license file reference in README.md
