### 2026-01-06 a2fa634
updated readme with license clarification
