### 2026-01-07 a0f8221
added license clarification to README
