### 2026-01-07 #9687 @david-gasinski
Hi Daniel, sorry for the delayed response. I've fixed the DOIs and updated the [licensing structure](https://github.com/Intelligent-Testing-Lab/cawsr/tree/110-joss-submission?tab=readme-ov-file#license) in our repository.

To answer your questions, during the development of CAWSR we forked and integrated some existing tooling into our code, hence the various licensing files. We ship everything as a unified package as all the components are tightly coupled, but we would only like code to be reviewed that is under our license.
In the repository we've added a new section [licensing](https://github.com/Intelligent-Testing-Lab/cawsr/tree/110-joss-submission?tab=readme-ov-file#license) to make it clear what tools we forked / modified, their licenses and what files belong under each license.
