### 2026-01-10 bd6b0ec
added autonomous_agent base class
