### 2026-01-11 3363c79
updated CARLA command for better reliability
