### 2026-01-11 3990ca7
added loop to spawn ego for redunancy
