### 2026-01-11 57b56c3
bug fix; exit loop on successful ego spawn
