### 2026-01-11 f1bfbd3
fixed bug with autoware double restarting
