### 2026-01-13 bc504e3
updated JOSS paper
