### 2026-01-15 5543156
updated JOSS paper to meet new reqs
