### 2026-01-15 #9687 @david-gasinski
Hi Daniel, the submission has been updated to meet the new requirements. The software to be reviewed as part of this submission is CAWSR, which is located in the following files in the repository:

```
srunner/scenarioconfigs/environment_configuration.py
srunner/scenario_decoder/json_to_xml_files.py
srunner/scenarioconfigs/carla_config.py
srunner/autoagents/autoware_agent.py
srunner/autoagents/autoware_nodes/
srunner/tools/results_manager.py
srunner/autoagents/agent_state/
srunner/objects/
Dockerfile
cawsr.py
srunner/tools/metrics_collector.py
srunner/tools/CARLA_manager.py
srunner/tools/log.py
srunner/scenariomanager/scenario_manager.py (modified)
srunner/autoagents/autoware_carla_interface/carla_autoware.py (modified)
srunner/autoagents/autoware_carla_interface/carla_ros.py (modified)
docker/autoware_msgs.tar.xz/src/autoware_cawsr_msgs
```
