### 2026-01-17 #9687 @danielskatz
I also think this meets JOSS's scope requirements, as best I can tell, but it's possible the editor or reviewers may have further comments or questions.
