### 2026-01-19 #9687 @hugoledoux
@david-gasinski 👋 I'll be the editor, and this is not my primary field, but I'll do my best.

Looking at the repo I saw this:
> Both CARLA and Autoware require a high-spec computer with a high-end Nvidia GPU. It is also possible to run a [distributed](https://github.com/Intelligent-Testing-Lab/cawsr/blob/main) setup with multiple machines to help ease the workload, or run the entire stack locally. Currently, only Linux is supported (guide was written on Ubuntu 24.04).
> Ensure the target machine(s) have the [Docker Engine](https://github.com/Intelligent-Testing-Lab/cawsr/blob/main) and [Nvidia Container toolkit](https://github.com/Intelligent-Testing-Lab/cawsr/blob/main) installed to enable gpu accelerated workflows in Docker.

which means that finding suitable reviewers might not be that easy... Can you please list here a few potential reviewers for this submission? Someone who obviously has no conflict of interests with any of the authors. If you list github usernames please remove the "@" so that they are not tagged and warned.
