### 2026-01-24 e1d29a1
fixed incorrect call to draw_line()
