### 2026-01-24 51abce8
fixed bug with incorrect call to draw_point()
