### 2026-01-24 09fea0d
adjust route visualiser line color to not interfere with traffic light detection
