+++ srunner/tools/route_visualisation.py
@@ -13,15 +13,15 @@ def visualise_route(route: list[carla.Location], lifetime=600):
 
     world_debug = CarlaDataProvider.get_world().debug
 
-    line_color = carla.Color(r=255, g=0, b=0)
-    waypoint_color = carla.Color(r=0, g=0, b=255)
+    line_color = carla.Color(r=0, g=0, b=60)
+    waypoint_color = carla.Color(r=0, g=0, b=80)
 
     for waypoint in range(len(route) - 1):
         begin = route[waypoint].location
         end = route[waypoint + 1].location
 
         world_debug.draw_line(
-            begin=begin, end=end, thickness=0.5, color=line_color, life_time=lifetime
+            begin=begin, end=end, thickness=0.02, color=line_color, life_time=lifetime
         )
         world_debug.draw_point(
             location=begin, size=0.1, color=waypoint_color, life_time=lifetime
@@ -29,5 +29,5 @@ def visualise_route(route: list[carla.Location], lifetime=600):
 
     # draw final waypoint
     world_debug.draw_point(
-        location=route[-1].location, size=0.1, color=waypoint_color, life_time=lifetime
+        location=route[-1].location, size=0.05, color=waypoint_color, life_time=lifetime
     )
