### 2026-01-31 9f3ceab
added code to force sensors to tick in carla at the same rate as autoware
