### 2026-01-31 ec16637
updated time ratio to track 1 / fixed_delta ticks
