### 2026-02-03 #9687 @david-gasinski
Hi Hugo, thank you for agreeing to edit our submission. Here is a list of potential reviewers who all have expertise in CARLA:
```
M M Abid Naziri∗ mnaziri@ncsu.edu North Carolina State University Raleigh, North Carolina, USA

Andrea Stocco†‡ andrea.stocco@tum.de Technical University of Munich, fortiss GmbH Munich, Bayern, Germany

Marcelo d’Amorim∗ mdamori@ncsu.edu North Carolina State University Raleigh, North Carolina, USA

Masoud Jamshidiyan Tehranimasoud.jamshidiyantehrani@usi.ch
Università della Svizzera italiana
Lugano, Switzerland

Jinhan Kimjinhan.kim@usi.ch
Università della Svizzera italiana
Lugano, Switzerland

Mingfei Chengmfcheng.2022@smu.edu.sg
Singapore Management University,
Singapore

Genjia Liu, LGJ1zed@sjtu.edu.cn
Shanghai Jiao Tong University, Shanghai, China.

LEI MA, University of Tokyo, Japan and University of Alberta, Canada (ma.lei [at] acm [dot] org)

Sandesh Bharadwaj
(from https://reviewers.joss.theoj.org/)
Github: tensorsofthewall
Email: sandesh.bharadwaj97@gmail.com
```
Note: I have suggested Sandesh Bharadwaj (tensorsofthewall) as a potential reviewer, as they possess experience in CARLA and have access to the required hardware. However, before realising the JOSS protocol, I had contacted them to confirm they had the minimum specs and provided a link to the repository. I am disclosing this to ensure full transparency, and will leave their participation to your discretion. I am completely open to selecting alternative reviewers if this prior contact is deemed a conflict of interest.
