### 2026-02-23 566ad8f
removed sensor tick; added wait for control command
