+++ srunner/autoagents/autoware_carla_interface/modules/carla_wrapper.py
@@ -218,9 +218,10 @@ class SensorWrapper(object):
                     yaw=sensor_spec["spawn_point"]["yaw"] + 0.0364,
                 )
 
-            bp.set_attribute(
-                "sensor_tick", str(round(1.0 / sensor_spec["publish_frequency"], 2))
-            )
+            # bp.set_attribute(
+            #    "sensor_tick", str(round(1.0 / sensor_spec["publish_frequency"], 2))
+            # )
+
             logger.info(
                 f"Sensor {sensor_spec['id']} tick rate set to {str(round(1.0 / sensor_spec['publish_frequency'], 2))}"
             )
