### 2026-02-23 bb6c278
added delay to ensure ticking remains constant
