### 2026-02-23 dc386b3
added more descriptive tick messages
