### 2026-02-25 #10113 @editorialbot
[REVIEW]: CAWSR: Carla-AutoWare Scenario Runner
**Submitting author:** @david-gasinski (<a href="http://orcid.org/0009-0008-7597-333X">David Gasinski</a>)
**Repository:** https://github.com/Intelligent-Testing-Lab/cawsr/
**Branch with paper.md** (empty if default branch): 110-joss-submission
**Version:** v1.0.0
