+++ srunner/tools/CARLA_manager.py
@@ -20,8 +20,9 @@ class CARLAManager(object):
     container_id = None
     port = 2000  # default
     fidelity = "Low"  # default
+    recording_dir = ""
     run_command = [
-        "docker run -dt --gpus all --net=host -v /tmp/.X11-unix:/tmp/.X11-unix:rw -e DISPLAY=$DISPLAY -e NVIDIA_DRIVER_CAPABILITIES=all -e XDG_RUNTIME_DIR=/tmp ghcr.io/intelligent-testing-lab/carla:0.9.15 ./CarlaUE4.sh -carla-rpc-port=2000 -quality-level=Low"
+        f"docker run -dt --gpus all --net=host -v /tmp/.X11-unix:/tmp/.X11-unix:rw -v {recording_dir}:/home/carla/recordings:rw -e DISPLAY=$DISPLAY -e NVIDIA_DRIVER_CAPABILITIES=all -e XDG_RUNTIME_DIR=/tmp ghcr.io/intelligent-testing-lab/carla:0.9.15 ./CarlaUE4.sh -carla-rpc-port=2000 -quality-level=Low"
     ]
 
     @staticmethod
@@ -32,6 +33,13 @@ class CARLAManager(object):
 
         CARLAManager.run_command = [
             f"docker run -dt --gpus all --net=host -v /tmp/.X11-unix:/tmp/.X11-unix:rw -e DISPLAY=$DISPLAY -e NVIDIA_DRIVER_CAPABILITIES=all -e XDG_RUNTIME_DIR=/tmp ghcr.io/intelligent-testing-lab/carla:0.9.15 ./CarlaUE4.sh -carla-rpc-port={CARLAManager.port} -quality-level={CARLAManager.fidelity}"
+        ]
+
+    @staticmethod
+    def _set_recording_dir(recording_dir: str) -> None:
+        CARLAManager.recording_dir = recording_dir
+        CARLAManager.run_command = [
+            f"docker run -dt --gpus all --net=host -v /tmp/.X11-unix:/tmp/.X11-unix:rw -v {CARLAManager.recording_dir}:/home/carla/recordings:rw -e DISPLAY=$DISPLAY -e NVIDIA_DRIVER_CAPABILITIES=all -e XDG_RUNTIME_DIR=/tmp ghcr.io/intelligent-testing-lab/carla:0.9.15 ./CarlaUE4.sh -carla-rpc-port={CARLAManager.port} -quality-level={CARLAManager.fidelity}"
         ]
 
     @staticmethod
