### 2026-02-25 2a455a6
(fix) moved CarlaState to carla_ros.py
