### 2026-02-25 78fab70
(fix) modified launch order of run_benchmark, fixed NoneType error with traffic manager
