### 2026-02-25 d977059
(fix) added error logs to starting CARLA container
