### 2026-02-25 03de090
(fix) add initialising state to not wait for control command
