### 2026-02-25 2b0d9a7
(fix) replaced Gametime.get_time calls with datetime.datetime.now
