### 2026-03-14 0694b10
added additional documentation to setup development environment
