+++ dev/pull_workspace.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+cd ..
+git clone https://github.com/Intelligent-Testing-Lab/cawsr_workspace.git
