### 2026-03-25 6aeb13b
(feat) updated autoware_carla_interface package to v0.50.0
