+++ srunner/autoagents/autoware_carla_interface/modules/sensor_kit_loader.py
@@ -1,0 +1,599 @@
+#!/usr/bin/env python3
+
+# Copyright 2024 Tier IV, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Sensor kit configuration loader for CARLA-Autoware interface."""
+
+import logging
+import os
+from pathlib import Path
+from typing import Any
+from typing import Dict
+from typing import List
+from typing import Optional
+
+from ament_index_python.packages import get_package_share_directory
+import yaml
+
+from srunner.autoagents.autoware_carla_interface.modules.coordinate_transformer import CoordinateTransformer
+from srunner.autoagents.autoware_carla_interface.modules.sensor_manager import SensorConfig
+
+# Default vehicle wheelbase in meters (distance between front and rear axles)
+DEFAULT_WHEELBASE = 2.850
+
+
+class SensorKitLoader:
+    """Loader for Autoware sensor kit configurations."""
+
+    def __init__(self, logger: Optional[logging.Logger] = None):
+        """Initialize sensor kit loader.
+
+        Args:
+            logger: Logger instance for output
+        """
+        self.logger = logger or logging.getLogger(__name__)
+        self.sensor_mapping: Dict[str, Any] = {}
+        self.sensor_kit_path: Optional[Path] = None
+        self.wheelbase: float = DEFAULT_WHEELBASE
+
+    def load_sensor_mapping(self, mapping_file: Optional[str] = None) -> bool:
+        """Load sensor mapping configuration.
+
+        Args:
+            mapping_file: Path to sensor mapping YAML file
+
+        Returns:
+            True if successfully loaded, False otherwise
+        """
+        try:
+            mapping_file = self._resolve_mapping_file_path(mapping_file)
+            if not mapping_file:
+                return False
+
+            with open(mapping_file, "r") as f:
+                self.sensor_mapping = yaml.safe_load(f)
+
+            self._validate_sensor_mapping_yaml()
+            self._load_vehicle_config()
+
+            self.logger.info(f"Loaded sensor mapping from: {mapping_file}")
+            return True
+
+        except ValueError as e:
+            self.logger.error(f"YAML validation error: {e}")
+            return False
+        except yaml.YAMLError as e:
+            self.logger.error(f"YAML parsing error in {mapping_file}: {e}")
+            return False
+        except Exception as e:
+            self.logger.error(f"Failed to load sensor mapping: {e}")
+            return False
+
+    def _resolve_mapping_file_path(self, mapping_file: Optional[str]) -> Optional[str]:
+        """Resolve sensor mapping file path.
+
+        Args:
+            mapping_file: Optional path to mapping file
+
+        Returns:
+            Resolved file path or None if not found
+        """
+        if not mapping_file:
+            pkg_dir = get_package_share_directory("autoware_carla_interface")
+            mapping_file = os.path.join(pkg_dir, "config", "sensor_mapping.yaml")
+
+        if os.path.exists(mapping_file):
+            return mapping_file
+
+        # Try source directory as fallback
+        source_mapping = Path(__file__).resolve().parents[3] / "config" / "sensor_mapping.yaml"
+        if source_mapping.exists():
+            return str(source_mapping)
+
+        self.logger.error(f"Sensor mapping file not found: {mapping_file}")
+        return None
+
+    def _validate_sensor_mapping_yaml(self):
[…]
