+++ srunner/autoagents/autoware_carla_interface/carla_ros.py
@@ -10,13 +10,11 @@
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
-# limitations under the License.sr/bin/env python
+# limitations under the License.
 
-import datetime
-import json
 import math
+import threading
 
-# pylint: disable=import-error
 from autoware_vehicle_msgs.msg import ControlModeReport
 from autoware_vehicle_msgs.msg import GearReport
 from autoware_vehicle_msgs.msg import SteeringReport
@@ -27,97 +25,64 @@ from cv_bridge import CvBridge
 from geometry_msgs.msg import Pose
 from geometry_msgs.msg import PoseWithCovarianceStamped
 import numpy
-import pathlib
+import rclpy
 from rosgraph_msgs.msg import Clock
 from sensor_msgs.msg import CameraInfo
-from sensor_msgs.msg import Image
 from sensor_msgs.msg import Imu
-from sensor_msgs.msg import PointCloud2
 from sensor_msgs.msg import PointField
 from std_msgs.msg import Header
 from tier4_vehicle_msgs.msg import ActuationCommandStamped
 from tier4_vehicle_msgs.msg import ActuationStatusStamped
 from transforms3d.euler import euler2quat
 
+# New modular sensor infrastructure
+from srunner.autoagents.autoware_carla_interface.modules import ROSPublisherManager
+from srunner.autoagents.autoware_carla_interface.modules import SensorKitLoader
+from srunner.autoagents.autoware_carla_interface.modules import SensorRegistry
 from srunner.scenariomanager.timer import GameTime
-from srunner.autoagents.autoware_carla_interface.modules.carla_utils import (
-    carla_location_to_ros_point,
-)
-from srunner.autoagents.autoware_carla_interface.modules.carla_utils import (
-    carla_rotation_to_ros_quaternion,
-)
+from srunner.autoagents.autoware_carla_interface.modules.carla_utils import carla_location_to_ros_point
+from srunner.autoagents.autoware_carla_interface.modules.carla_utils import carla_rotation_to_ros_quaternion
 from srunner.autoagents.autoware_carla_interface.modules.carla_utils import create_cloud
-from srunner.autoagents.autoware_carla_interface.modules.carla_utils import (
-    ros_pose_to_carla_transform,
-)
-from srunner.autoagents.autoware_carla_interface.modules.carla_wrapper import (
-    SensorInterface,
-)
-from srunner.tools.CARLA_manager import CARLAManager
+from srunner.autoagents.autoware_carla_interface.modules.carla_utils import ros_pose_to_carla_transform
+from srunner.autoagents.autoware_carla_interface.modules.carla_wrapper import SensorInterface
 
 
 class carla_ros2_interface(object):
-    def __init__(self, node):
-        self.sensor_interface = SensorInterface()
-        self.prev_timestamp = None
-        self.prev_steer_output = 0.0
-        self.tau = 0.2
-        self.timestamp = None
-        self.ego_actor = None
-        self.physics_control = None
-        self.channels = 0
-        self.id_to_sensor_type_map = {}
-        self.id_to_camera_info_map = {}
-        self.cv_bridge = CvBridge()
-        self.first_ = True
-        self.pub_lidar = {}
-        self.sensor_frequencies = {
-            "top": 11,
-            "left": 11,
-            "right": 11,
-            "camera": 11,
-            "imu": 50,
-            "status": 50,
-            "pose": 2,
-        }
-        # self.publish_prev_times = {
-        #    sensor: GameTime.get_time() for sensor in self.sensor_frequencies
-        # }
 
[…]
