+++ srunner/autoagents/autoware_carla_interface/carla_ros.py
@@ -40,20 +40,33 @@ from srunner.autoagents.autoware_carla_interface.modules import ROSPublisherMana
 from srunner.autoagents.autoware_carla_interface.modules import SensorKitLoader
 from srunner.autoagents.autoware_carla_interface.modules import SensorRegistry
 from srunner.scenariomanager.timer import GameTime
-from srunner.autoagents.autoware_carla_interface.modules.carla_utils import carla_location_to_ros_point
-from srunner.autoagents.autoware_carla_interface.modules.carla_utils import carla_rotation_to_ros_quaternion
+from srunner.autoagents.autoware_carla_interface.modules.carla_utils import (
+    carla_location_to_ros_point,
+)
+from srunner.autoagents.autoware_carla_interface.modules.carla_utils import (
+    carla_rotation_to_ros_quaternion,
+)
 from srunner.autoagents.autoware_carla_interface.modules.carla_utils import create_cloud
-from srunner.autoagents.autoware_carla_interface.modules.carla_utils import ros_pose_to_carla_transform
-from srunner.autoagents.autoware_carla_interface.modules.carla_wrapper import SensorInterface
+from srunner.autoagents.autoware_carla_interface.modules.carla_utils import (
+    ros_pose_to_carla_transform,
+)
+from srunner.autoagents.autoware_carla_interface.modules.carla_wrapper import (
+    SensorInterface,
+)
 
 
 class carla_ros2_interface(object):
-
     def _initialize_parameters(self):
         """Initialize and declare ROS 2 parameters."""
-        self.parameters = {            
-            "sensor_kit_name": (rclpy.Parameter.Type.STRING, "carla_sensor_kit_description"),
-            "sensor_mapping_file": (rclpy.Parameter.Type.STRING, "/autoware_scenario_runner/srunner/autoagents/autoware_carla_interface"),
+        self.parameters = {
+            "sensor_kit_name": (
+                rclpy.Parameter.Type.STRING,
+                "carla_sensor_kit_description",
+            ),
+            "sensor_mapping_file": (
+                rclpy.Parameter.Type.STRING,
+                "/autoware_scenario_runner/srunner/autoagents/autoware_carla_interface/config/sensor_mapping.yaml",
+            ),
         }
 
         self.param_values = {}
@@ -63,7 +76,9 @@ class carla_ros2_interface(object):
                 self.ros2_node.declare_parameter(param_name, default_value)
             else:
                 self.ros2_node.declare_parameter(param_name, param_type)
-            self.param_values[param_name] = self.ros2_node.get_parameter(param_name).value
+            self.param_values[param_name] = self.ros2_node.get_parameter(
+                param_name
+            ).value
 
     def _initialize_clock_publisher(self):
         """Initialize and publish initial clock message."""
@@ -102,7 +117,10 @@ class carla_ros2_interface(object):
     def _initialize_subscriptions(self):
         """Initialize all ROS 2 subscriptions."""
         self.sub_control = self.ros2_node.create_subscription(
-            ActuationCommandStamped, "/control/command/actuation_cmd", self.control_callback, 1
+            ActuationCommandStamped,
+            "/control/command/actuation_cmd",
+            self.control_callback,
+            1,
         )
         self.sub_vehicle_initialpose = self.ros2_node.create_subscription(
             PoseWithCovarianceStamped, "initialpose", self.initialpose_callback, 1
@@ -148,11 +166,15 @@ class carla_ros2_interface(object):
         if param_value:
             return param_value
 
-        mapping_default = self.sensor_loader.sensor_mapping.get("default_sensor_kit_name", "")
+        mapping_default = self.sensor_loader.sensor_mapping.get(
+            "default_sensor_kit_name", ""
+        )
[…]
