### 2026-04-16 5eae87f
(feat) added background behaviour to scenario definition; (bug) fixed incorrect timing sleep logic in scenario manager
