### 2026-04-17 05fbf86
(feat) added topic to publish CAWSR state
