### 2026-04-22 6ee91ae
(feat) added additional camera publishing for yolo-based object detection
