+++ cawsr.py
@@ -67,8 +67,12 @@ class AWScenarioRunner(object):
     aw_agent = None  # autoware agent
 
     def __init__(self, cawsr_config: dict, carla_conf: CARLA) -> None:
-        """
-        Setup Scenario Manager and the Carla client
+        """Initialises CAWSR, loading the Autoware agent module and setting up signal handlers for cleanup on shutdown.
+        Also initialises the ScenarioDefinitionManager for managing results directories and files.
+
+        Args:
+            cawsr_config (dict): CAWSR configuration options loaded from YAML file
+            carla_conf (CARLA): CARLA-specific configuration options loaded from YAML file
         """
         self.DEBUG = cawsr_config["debug"]
         self._conf = cawsr_config
@@ -120,6 +124,18 @@ class AWScenarioRunner(object):
         current_definiton: Union[dict, None],
         algorithm: Union[Callable, None],
     ) -> None:
+        """Starts the CAWSR scenario based on the provided configuration and seed
+        Handles connection to the Autoware agent and CARLA client, and runs the scenario manager until the scenario is complete or a shutdown signal is received.
+        Args:
+            route_config (RouteScenarioConfiguration): RouteSCenarioConfiguration object configuration
+            env_config (EnvironmentConfig): EnvironmentConfig object containing the environment configuration
+            scenario_name (str): Name of the scenario to run
+            seed (int): Random seed for reproducibility
+            algorithm_mode (bool): Whether to load algorithmic scenario generation loop, or just execute scenarios in sequence
+            result_ (_type_): Result object for storing scenario results
+            current_definiton (Union[dict, None]): Optional dict containing the current scenario definition, used for algorithmic scenario generation
+            algorithm (Union[Callable, None]): Optional Callable algorithm class, used for algorithmic scenario generation. See docs for expected interface.
+        """
         logger.info("Initialising Scenario Manager...")
         self.scenario_manager = ScenarioManager(
             self.DEBUG,
@@ -200,18 +216,14 @@ class AWScenarioRunner(object):
 
         logger.info("Loading route...")
 
-        # TO DO
-        # interpolate route at a larger distance (i.e 5m) to reduce waypoints
-        # remove segment sampling from awagent.set_route
         gps_route, route = route_manipulation.interpolate_trajectory(
             route_config.keypoints
         )
         route_config.agent.set_global_plan(gps_route, route)  # set agent route
 
-        # visualise the route in CARLA
+        # visualise the route in CARLA and set location to first waypoint
         visualise_route([waypoint[0] for waypoint in route])
-
-        ego.prepare_ego(route[0][0])  # set location to first waypoint
+        ego.prepare_ego(route[0][0])
 
         # allow the agent X ticks to initialize sensors and set the route
         logger.info("Initialising agent route...")
@@ -265,6 +277,7 @@ class AWScenarioRunner(object):
             result = False
 
         self.carla_client.stop_recorder()
+
         # stop the MetricsCollector thread
         MetricsCollector.reset()
 
@@ -292,6 +305,7 @@ class AWScenarioRunner(object):
         result_.put(result_dict)
 
     def _tick_carla(self) -> None:
+        """Advances CARLA 1 tick into the future"""
         timestamp = None
         world = CarlaDataProvider.get_world()
         if world:
@@ -323,9 +337,11 @@ class AWScenarioRunner(object):
         )
 
     def run_algorithm(self) -> None:
-        """Executes CAWSR in algorithm mode"""
+        """Executes CAWSR in algorithm mode, loading a custom algorithm interface for scenario generation and optimisation.
+        Scenarios are generated iteratively based on the previous scenario's performance, and the initial scenario definition provided.
+        """
 
-        # load the algorithm and scenario defintion
[…]
