### 2026-04-23 89c97db
updated publishers for autoware camera topics
