+++ srunner/tools/results_manager.py
@@ -73,7 +73,9 @@ class ScenarioDefinitionManager(object):
 
         # make the recordings folder
         os.makedirs(os.path.join(full_path, "recording"), exist_ok=True)
-        self.recording_path = Path(os.path.join(full_path, "recording")).resolve()
+        self.recording_path = (
+            Path(os.path.join(full_path, "recording")).resolve().as_posix()
+        )
 
         os.makedirs(full_path, exist_ok=True)  # exist_ok=True, no need to error handle
         self.last_scenario = full_path
