### 2026-04-27 4970a58
updated dev workspace script
