### 2026-04-27 7bd1afe
(feat) added sensor configuration to scenario definition
