+++ docs/DEVELOPMENT.md
@@ -20,10 +20,10 @@ This will be mounted into the dev container to allow you to configure CAWSR as u
 ## Running CAWSR
 The `dev/docker-compose.yml` file includes the services needed to launch the development container. Ensure the file is in the root of the repository and run:
 ```bash
-docker-compose -f dev/docker-compose.yml up cawsr autoware-latest
+docker compose -f dev/docker-compose.yml up cawsr autoware-latest
 ```
 
-Development follows a similar pattern to the operation of CAWSR - simply start the autoware container and CAWSR. The `cawsr_dev` image mounts root of the repository into the container, so any changes you make are applied in CAWSR.
+Development follows a similar pattern to the operation of CAWSR - simply start the autoware container and CAWSR. The `cawsr_dev` image mounts root of the repository into the container, so any changes you make are applied in CAWSR. If you need to configure CAWSR, either edit `dev/.env` to change the runtime configuration, or `configs/` to configure CAWSR.
 
 ## Contributing
 See [CONTRIBUTING.md](../CONTRIBUTING.md)
