### 2026-04-27 b9a4d21
(bug) fixed division by zero bug
