### 2026-04-28 #10113 @nitishsanghi
@hugoledoux the paper is still missing 'State of the field'. Can we request the authors to update the paper to include that section.
