### 2026-05-05 2568ca9
added missing leaderboard scenario definition
