### 2026-05-05 de651cc
fixed bug with scenario process crashing and causing cascading failure
