### 2026-05-06 b60a199
updated logic to shutdown and restart autoware every scenario iteration
