+++ srunner/autoagents/autoware_agent.py
@@ -59,9 +59,7 @@ class AutowareAgent(AutonomousAgent):
         self.carla_interface = InitializeInterface(self.config, self._node)
 
         self.state_node = state_node.StateNode(self.autoware_state, self._node_state)
-        if config.iteration == 1:
-            # only reset autoware on the first run
-            self.state_node.reset_autoware(self.config.town, self.config.ego_name)
+        self.state_node.reset_autoware(self.config.town, self.config.ego_name)
 
         self.route_node = route_node.RouteNode(self.autoware_state, self._node_state)
         self.autoware_node = autoware_node.AutowareNode(
@@ -123,7 +121,7 @@ class AutowareAgent(AutonomousAgent):
         )
 
     def cleanup(self) -> None:
-        self.destroy()
+        self.destroy(cleanup=True)
 
     def destroy(self, cleanup=False) -> None:
         """Cleanup"""
