### 2026-05-09 80720df
(refactor) updated get_data to be frame aware; CAWSR now waits for sensor callbacks to fire before proceeding with sensor publishing, ensuring sensor data is latest frame. This allows the migration from wall time (maximum wall:sim ratio is 1:1) to game_time
