### 2026-05-09 9c08dcf
added sim vs wall ratio logging
