### 2026-05-09 bbd4531
(fix) attempted performance fixes to reduce sim:wall ratio
