### 2026-05-09 89a4c0e
(bug) fixed issue with incorrect init_loop() before causing localisation issue
