### 2026-05-09 ffa11a8
(fix) fixed bug with _tick_carla() calling agent class before initialisation
