### 2026-05-09 1ed5643
(fix) moved autoware_restart to allow time for ROS callbacks to spin up
