### 2026-05-09 449edcc
(fix) attempted fix for localistaion issue during init loop. Merged run_init() and run_step(), controlling initialisation via a flag.
