### 2026-05-09 94a38d1
(fix) revert stale sensor data check, testing if this could be the cause
