### 2026-05-09 438ca4c
(bug) added last_control variable to apply the last received control if autoware hasn't ticket yet
