### 2026-05-09 8b3f32c
added camera smoothing
