+++ srunner/scenariomanager/scenario_manager.py
@@ -112,6 +112,8 @@ class ScenarioManager(object):
             self.world_cam = CarlaDataProvider.get_world().get_spectator()
             self._camera_offset = carla.Location(x=0, y=0, z=50)
             self._camera_pitch = -90.0  # degrees
+            self._smooth_cam_loc = None
+            self._smooth_cam_yaw = None
 
     def run_scenario(self):
         """
@@ -217,12 +219,24 @@ class ScenarioManager(object):
     def _tick_spectator_cam(self, ego: carla.Actor) -> None:
         """Ticks the spectator camera for the chosen ego"""
         vehicle_transform = ego.get_transform()
-        delta_spec_loc = vehicle_transform.location + self._camera_offset
+        target_loc = vehicle_transform.location + self._camera_offset
+        target_yaw = vehicle_transform.rotation.yaw
+
+        smoothing = 0.3
+
+        if self._smooth_cam_loc is None:
+            self._smooth_cam_loc = target_loc
+            self._smooth_cam_yaw = target_yaw
+        else:
+            self._smooth_cam_loc.x += (target_loc.x - self._smooth_cam_loc.x) * smoothing
+            self._smooth_cam_loc.y += (target_loc.y - self._smooth_cam_loc.y) * smoothing
+            self._smooth_cam_loc.z += (target_loc.z - self._smooth_cam_loc.z) * smoothing
+            self._smooth_cam_yaw += (target_yaw - self._smooth_cam_yaw) * smoothing
 
         delta_spec_trans = carla.Transform(
-            delta_spec_loc,
+            self._smooth_cam_loc,
             carla.Rotation(
-                pitch=self._camera_pitch, yaw=vehicle_transform.rotation.yaw, roll=0
+                pitch=self._camera_pitch, yaw=self._smooth_cam_yaw, roll=0
             ),
         )
         self.world_cam.set_transform(delta_spec_trans)
