### 2026-05-10 d2b5139
made camera smoothing more aggressive
