### 2026-05-10 28e69cb
refactor of LiDaR msg structuring to improve performance
