+++ srunner/autoagents/autoware_carla_interface/carla_ros.py
@@ -217,60 +217,40 @@ class carla_ros2_interface(object):
         header = self.get_msg_header(
             frame_id="velodyne_top_changed", timestamp=timestamp
         )
+
+        lidar_data = numpy.frombuffer(
+            carla_lidar_measurement.raw_data, dtype=numpy.float32
+        ).reshape(-1, 4)
+
+        point_count = lidar_data.shape[0]
+        buf = numpy.empty(point_count, dtype=[
+            ("x", "f4"), ("y", "f4"), ("z", "f4"),
+            ("intensity", "u1"), ("return_type", "u1"), ("channel", "u2"),
+        ])
+
+        buf["x"] = lidar_data[:, 0]
+        buf["y"] = -lidar_data[:, 1]
+        buf["z"] = lidar_data[:, 2]
+        buf["intensity"] = numpy.clip(lidar_data[:, 3], 0, 1) * 255
+        buf["return_type"] = 0
+
+        channels = self.sensors["sensors"][1]["channels"]
+        ch_starts = [0]
+        for i in range(channels):
+            ch_starts.append(ch_starts[-1] + carla_lidar_measurement.get_point_count(i))
+        for i in range(channels):
+            buf["channel"][ch_starts[i] : ch_starts[i + 1]] = i
+
         fields = [
             PointField(name="x", offset=0, datatype=PointField.FLOAT32, count=1),
             PointField(name="y", offset=4, datatype=PointField.FLOAT32, count=1),
             PointField(name="z", offset=8, datatype=PointField.FLOAT32, count=1),
             PointField(name="intensity", offset=12, datatype=PointField.UINT8, count=1),
-            PointField(
-                name="return_type", offset=13, datatype=PointField.UINT8, count=1
-            ),
+            PointField(name="return_type", offset=13, datatype=PointField.UINT8, count=1),
             PointField(name="channel", offset=14, datatype=PointField.UINT16, count=1),
         ]
 
-        lidar_data = numpy.frombuffer(
-            carla_lidar_measurement.raw_data, dtype=numpy.float32
-        ).reshape(-1, 4)
-        intensity = lidar_data[:, 3]
-        intensity = (
-            numpy.clip(intensity, 0, 1) * 255
-        )  # CARLA lidar intensity values are between 0 and 1
-        intensity = intensity.astype(numpy.uint8).reshape(-1, 1)
-
-        return_type = numpy.zeros((lidar_data.shape[0], 1), dtype=numpy.uint8)
-        channel = numpy.empty((0, 1), dtype=numpy.uint16)
-        self.channels = self.sensors["sensors"]
-
-        for i in range(self.channels[1]["channels"]):
-            current_ring_points_count = carla_lidar_measurement.get_point_count(i)
-            channel = numpy.vstack(
-                (
-                    channel,
-                    numpy.full((current_ring_points_count, 1), i, dtype=numpy.uint16),
-                )
-            )
-
-        lidar_data = numpy.hstack((lidar_data[:, :3], intensity, return_type, channel))
-        lidar_data[:, 1] *= -1
-
-        dtype = [
-            ("x", "f4"),
-            ("y", "f4"),
-            ("z", "f4"),
-            ("intensity", "u1"),
-            ("return_type", "u1"),
-            ("channel", "u2"),
-        ]
-
-        structured_lidar_data = numpy.zeros(lidar_data.shape[0], dtype=dtype)
-        structured_lidar_data["x"] = lidar_data[:, 0]
-        structured_lidar_data["y"] = lidar_data[:, 1]
-        structured_lidar_data["z"] = lidar_data[:, 2]
[…]
