### 2026-05-10 57d12b7
(fix) performance improvements; removed unnecessary time.sleep(0) calls for GIL yield; updated carlaDataProvider to merge 3 update loops into 1; reduced RPC calls by ticking spectator cam at 4HZ instead of 20
