+++ srunner/scenariomanager/scenario_manager.py
@@ -62,6 +62,7 @@ class ScenarioManager(object):
 
         self._running = False
         self._timestamp_last_run = 0.0
+        self._cam_tick = 0
         self.scenario_duration_system = 0.0
         self.scenario_duration_game = 0.0
         self.start_system_time = None
@@ -73,6 +74,7 @@ class ScenarioManager(object):
         """
         self._running = False
         self._timestamp_last_run = 0.0
+        self._cam_tick = 0
         self.scenario_duration_system = 0.0
         self.scenario_duration_game = 0.0
         self.start_system_time = None
@@ -137,10 +139,6 @@ class ScenarioManager(object):
             world = CarlaDataProvider.get_world()
             if world:
                 world.tick()
-                # Yield the GIL several times so CARLA internal threads
-                # can push sensor callbacks into the data buffers.
-                for _ in range(5):
-                    time.sleep(0)
             MetricsCollector.update_key(
                 "carla_time", (time.perf_counter_ns() / 1e6) - _tick_carla_start
             )
@@ -211,7 +209,9 @@ class ScenarioManager(object):
             )
 
             if self.follow_ego:
-                self._tick_spectator_cam(self.ego_vehicles[0])  # type: ignore
+                self._cam_tick += 1
+                if self._cam_tick % 5 == 0:
+                    self._tick_spectator_cam(self.ego_vehicles[0])  # type: ignore
 
             if self.scenario_tree.status != py_trees.common.Status.RUNNING:
                 self._running = False
