### 2026-05-10 612e477
(revert) reverted changes to CarlaDataProvider and spectator camera, not much impact on performance
