### 2026-05-10 39aee43
(bug) added pitch camera smoothing; fixed call to uninitialised class attribute scenario_manager in destroy(); added silent_spin func to suppress ros2 executor shutdown
