### 2026-05-10 92aa426
control loss fix, AV_Control in pytrees was never set
