### 2026-05-10 d7434a4
(feat) added functionality to show scenario triggers in CARLA
