+++ srunner/scenarios/route_scenario.py
@@ -269,28 +269,10 @@ class RouteScenario(BasicScenario):
         """
         all_scenario_classes = self.get_all_scenario_classes()
         self.list_scenarios = []
+        self.trigger_markers = []
         ego_data = ActorConfigurationData(
             ego_vehicle.type_id, ego_vehicle.get_transform(), "hero"
         )
-
-        if debug:
-            tmap = CarlaDataProvider.get_map()
-            for scenario_config in scenario_definitions:
-                scenario_loc = scenario_config.trigger_points[0].location
-                debug_loc = tmap.get_waypoint(
-                    scenario_loc
-                ).transform.location + carla.Location(z=0.2)
-                world.debug.draw_point(
-                    debug_loc, size=0.2, color=carla.Color(128, 0, 0), life_time=timeout
-                )
-                world.debug.draw_string(
-                    debug_loc,
-                    str(scenario_config.name),
-                    draw_shadow=False,
-                    color=carla.Color(0, 0, 128),
-                    life_time=timeout,
-                    persistent_lines=True,
-                )
 
         for scenario_number, scenario_config in enumerate(scenario_definitions):
             scenario_config.ego_vehicles = [ego_data]
@@ -321,6 +303,14 @@ class RouteScenario(BasicScenario):
                 continue
 
             self.list_scenarios.append(scenario_instance)
+
+        for scenario_config in scenario_definitions:
+            trigger = scenario_config.trigger_points[0]
+            self.trigger_markers.append({
+                "name": scenario_config.name,
+                "location": trigger.location + carla.Location(z=1.0),
+                "route_var_name": scenario_config.route_var_name,
+            })
 
     # pylint: enable=no-self-use
     def _initialize_actors(self, config):
