### 2026-05-10 d4a8866
(fix) changed camera following algorithm to use ego forward vector instead of rotation.yaw
