### 2026-05-10 55f8992
(fix) fixed issue with trigger box not showing in CARLA
