### 2026-05-10 5fa7093
added logging to inform when scenario event starts
