+++ srunner/scenariomanager/scenarioatomics/atomic_behaviors.py
@@ -4985,7 +4985,7 @@ class ScenarioTriggerer(AtomicBehavior):
 
         # Check which scenarios can be triggered
         blackboard = py_trees.blackboard.Blackboard()
-        for black_var_name, scen_location in self._blackboard_list:
+        for black_var_name, scen_location, scen_name in self._blackboard_list:
             # Close enough
             scen_distance = route_location.distance(scen_location)
             condition1 = bool(scen_distance < self._distance)
@@ -5000,6 +5000,8 @@ class ScenarioTriggerer(AtomicBehavior):
             if condition1 and condition2 and condition3:
                 _ = blackboard.set(black_var_name, True)
                 self._triggered_scenarios.append(black_var_name)
+
+                print(f"Scenario '{scen_name}' triggered")
 
                 if self._debug:
                     self._world.debug.draw_point(
