### 2026-05-10 9c39619
(feat) added HUD logging to show current control and variation
