### 2026-05-10 1a32081
(fix) updated camera smoothing; moved HUD to top left corner
