### 2026-05-10 01a146e
reverted camera smoothing - still jittering. Instead camera now follows ego with a fixed yaw
