+++ srunner/scenariomanager/scenario_manager.py
@@ -111,10 +111,8 @@ class ScenarioManager(object):
 
         if follow_ego:
             self.world_cam = CarlaDataProvider.get_world().get_spectator()
-            self._camera_offset = carla.Location(x=0, y=0, z=50)
+            self._camera_offset = carla.Location(x=0, y=0, z=70)
             self._camera_pitch = -90.0  # degrees
-            self._smooth_cam_loc = None
-            self._smooth_cam_yaw = 0.0
             self._cam_transform = None
 
         self._draw_trigger_markers()
@@ -219,37 +217,10 @@ class ScenarioManager(object):
 
     def _tick_spectator_cam(self, ego: carla.Actor) -> None:
         """Ticks the spectator camera for the chosen ego"""
-        vehicle_transform = ego.get_transform()
-        target_loc = vehicle_transform.location + self._camera_offset
-        forward = vehicle_transform.get_forward_vector()
-        target_yaw = math.degrees(math.atan2(forward.y, forward.x))
-
-        smoothing_loc = 0.05
-        smoothing_rot = 0.02
-
-        if self._smooth_cam_loc is None:
-            self._smooth_cam_loc = target_loc
-            self._smooth_cam_yaw = target_yaw
-        else:
-            self._smooth_cam_loc.x += (
-                target_loc.x - self._smooth_cam_loc.x
-            ) * smoothing_loc
-            self._smooth_cam_loc.y += (
-                target_loc.y - self._smooth_cam_loc.y
-            ) * smoothing_loc
-            self._smooth_cam_loc.z += (
-                target_loc.z - self._smooth_cam_loc.z
-            ) * smoothing_loc
-
-            vel = ego.get_velocity()
-            speed = math.sqrt(vel.x ** 2 + vel.y ** 2)
-            if speed > 0.5:
-                dyaw = (target_yaw - self._smooth_cam_yaw + 180.0) % 360.0 - 180.0
-                self._smooth_cam_yaw = (self._smooth_cam_yaw + dyaw * smoothing_rot) % 360.0
-
+        ego_trans = ego.get_transform()
         self._cam_transform = carla.Transform(
-            self._smooth_cam_loc,
-            carla.Rotation(pitch=self._camera_pitch, yaw=self._smooth_cam_yaw, roll=0),
+            ego_trans.location + self._camera_offset,
+            carla.Rotation(pitch=self._camera_pitch, yaw=0.0, roll=0.0),
         )
         self.world_cam.set_transform(self._cam_transform)
 
@@ -281,7 +252,7 @@ class ScenarioManager(object):
             )
 
     def _draw_control_hud(self, ego):
-        if self._smooth_cam_loc is None:
+        if self._cam_transform is None:
             return
         world = CarlaDataProvider.get_world()
         if world is None:
