### 2026-05-11 95acdaf
(bug) attempted fixes to help with route setting on corners; increased yaw precision sent to autowar
