+++ srunner/autoagents/autoware_agent.py
@@ -38,6 +38,8 @@ def _silent_spin(executor):
 
 
 class AutowareAgent(AutonomousAgent):
+    MAX_ROUTE_RETRIES = 3
+
     timestamp = None
     agent_set_route = False
     scenario_loaded = False
@@ -99,7 +101,8 @@ class AutowareAgent(AutonomousAgent):
             self.autoware_state, self._node_state
         )
 
-        self.sent_route = False
+        self._reset_route_state()
+        self._route_retry_count = 0
 
         self.carla_interface.load_world()
         self.carla_interface.run_bridge()
@@ -118,6 +121,21 @@ class AutowareAgent(AutonomousAgent):
 
         logger.info("Clearing route...")
         self.route_node.request_clear_route()
+
+    def _reset_route_state(self):
+        self.sent_route = False
+        self._route_was_calculating = False
+
+    def _retry_route(self):
+        if self._route_retry_count < self.MAX_ROUTE_RETRIES:
+            self._route_retry_count += 1
+            self._reset_route_state()
+            logger.info("Clearing route before retry...")
+            self.route_node.request_clear_route()
+        else:
+            logger.error(
+                f"Route setting failed after {self.MAX_ROUTE_RETRIES} attempts."
+            )
 
     def _convert_to_waypoint(self, point):
         """Returns a waypoint
@@ -186,6 +204,22 @@ class AutowareAgent(AutonomousAgent):
 
             self.route_node.publish_route(goal_pose, waypoints)
             self.sent_route = True
+            logger.info("Route published to Autoware")
+
+        if self.sent_route:
+            if self.autoware_state.is_planning():
+                self._route_was_calculating = True
+
+            if self.autoware_state.route_set():
+                self._route_retry_count = 0
+                self._route_was_calculating = False
+
+            elif self.autoware_state.route_rejected(self._route_was_calculating):
+                logger.warning(
+                    "Route calculation failed: state transitioned 3->2. "
+                    f"Retry {self._route_retry_count + 1}/{self.MAX_ROUTE_RETRIES}"
+                )
+                self._retry_route()
 
         if (
             self.scenario_loaded
