### 2026-05-11 60ca418
(bug) update route messages to use same timestamp; updated route calculation logic
