### 2026-05-12 576c060
(bug) minor fixes to improve scenarios
