+++ srunner/tools/scenario_helper.py
@@ -38,7 +38,6 @@ def get_distance_along_route(route, target_location):
     target_location_from_wp = wmap.get_waypoint(target_location).transform.location
 
     for position, _ in route:
-
         location = target_location_from_wp
 
         # Don't perform any calculations for the first route point
@@ -47,8 +46,12 @@ def get_distance_along_route(route, target_location):
             continue
 
         # Calculate distance between previous and current route point
-        interval_length_squared = ((prev_position.x - position.x) ** 2) + ((prev_position.y - position.y) ** 2)
-        distance_squared = ((location.x - prev_position.x) ** 2) + ((location.y - prev_position.y) ** 2)
+        interval_length_squared = ((prev_position.x - position.x) ** 2) + (
+            (prev_position.y - position.y) ** 2
+        )
+        distance_squared = ((location.x - prev_position.x) ** 2) + (
+            (location.y - prev_position.y) ** 2
+        )
 
         # Close to the current position? Stop calculation
         if distance_squared < 0.01:
@@ -62,7 +65,8 @@ def get_distance_along_route(route, target_location):
             while wp is not None:
                 new_location = wp.transform.location
                 new_distance_squared = ((new_location.x - prev_position.x) ** 2) + (
-                    (new_location.y - prev_position.y) ** 2)
+                    (new_location.y - prev_position.y) ** 2
+                )
 
                 if np.sign(starting_wp.lane_id) != np.sign(wp.lane_id):
                     break
@@ -79,7 +83,8 @@ def get_distance_along_route(route, target_location):
             while wp is not None:
                 new_location = wp.transform.location
                 new_distance_squared = ((new_location.x - prev_position.x) ** 2) + (
-                    (new_location.y - prev_position.y) ** 2)
+                    (new_location.y - prev_position.y) ** 2
+                )
 
                 if np.sign(starting_wp.lane_id) != np.sign(wp.lane_id):
                     break
@@ -105,8 +110,9 @@ def get_distance_along_route(route, target_location):
             if prev_wp and curr_wp and wp:
                 if wp.road_id in (prev_wp.road_id, curr_wp.road_id):
                     # Roads match, now compare the sign of the lane ids
-                    if (np.sign(wp.lane_id) == np.sign(prev_wp.lane_id) or
-                            np.sign(wp.lane_id) == np.sign(curr_wp.lane_id)):
+                    if np.sign(wp.lane_id) == np.sign(prev_wp.lane_id) or np.sign(
+                        wp.lane_id
+                    ) == np.sign(curr_wp.lane_id):
                         # The location is within the current route interval
                         covered_distance += math.sqrt(distance_squared)
                         found = True
@@ -129,13 +135,18 @@ def get_crossing_point(actor):
     while not wp_cross.is_intersection:
         wp_cross = wp_cross.next(2)[0]
 
-    crossing = carla.Location(x=wp_cross.transform.location.x,
-                              y=wp_cross.transform.location.y, z=wp_cross.transform.location.z)
+    crossing = carla.Location(
+        x=wp_cross.transform.location.x,
+        y=wp_cross.transform.location.y,
+        z=wp_cross.transform.location.z,
+    )
 
     return crossing
 
 
-def get_geometric_linear_intersection(ego_location, other_location, move_to_junction=False):
+def get_geometric_linear_intersection(
+    ego_location, other_location, move_to_junction=False
+):
     """
     Obtain a intersection point between two actor's location by using their waypoints (wp)
 
[…]
