+++ srunner/scenarios/signalized_junction_right_turn.py
@@ -16,30 +16,43 @@ import py_trees
 import carla
 
 from srunner.scenariomanager.carla_data_provider import CarlaDataProvider
-from srunner.scenariomanager.scenarioatomics.atomic_behaviors import ActorFlow, TrafficLightFreezer, ScenarioTimeout
-from srunner.scenariomanager.scenarioatomics.atomic_criteria import CollisionTest, ScenarioTimeoutTest
-from srunner.scenariomanager.scenarioatomics.atomic_trigger_conditions import WaitEndIntersection, DriveDistance
+from srunner.scenariomanager.scenarioatomics.atomic_behaviors import (
+    ActorFlow,
+    TrafficLightFreezer,
+    ScenarioTimeout,
+)
+from srunner.scenariomanager.scenarioatomics.atomic_criteria import (
+    CollisionTest,
+    ScenarioTimeoutTest,
+)
+from srunner.scenariomanager.scenarioatomics.atomic_trigger_conditions import (
+    WaitEndIntersection,
+    DriveDistance,
+)
 from srunner.scenarios.basic_scenario import BasicScenario
-from srunner.tools.scenario_helper import (generate_target_waypoint,
-                                           get_junction_topology,
-                                           filter_junction_wp_direction,
-                                           get_same_dir_lanes,
-                                           get_closest_traffic_light)
+from srunner.tools.scenario_helper import (
+    generate_target_waypoint,
+    get_junction_topology,
+    filter_junction_wp_direction,
+    get_same_dir_lanes,
+    get_closest_traffic_light,
+)
 
 from srunner.tools.background_manager import HandleJunctionScenario
 
 
 def get_value_parameter(config, name, p_type, default):
     if name in config.other_parameters:
-        return p_type(config.other_parameters[name]['value'])
+        return p_type(config.other_parameters[name]["value"])
     else:
         return default
+
 
 def get_interval_parameter(config, name, p_type, default):
     if name in config.other_parameters:
         return [
-            p_type(config.other_parameters[name]['from']),
-            p_type(config.other_parameters[name]['to'])
+            p_type(config.other_parameters[name]["from"]),
+            p_type(config.other_parameters[name]["to"]),
         ]
     else:
         return default
@@ -51,8 +64,16 @@ class JunctionRightTurn(BasicScenario):
     colliding with vehicles coming from its left
     """
 
-    def __init__(self, world, ego_vehicles, config, randomize=False, debug_mode=False, criteria_enable=True,
-                 timeout=80):
+    def __init__(
+        self,
+        world,
+        ego_vehicles,
+        config,
+        randomize=False,
+        debug_mode=False,
+        criteria_enable=True,
+        timeout=80,
+    ):
         """
         Setup all relevant parameters and create scenario
         """
@@ -60,28 +81,34 @@ class JunctionRightTurn(BasicScenario):
         self._map = CarlaDataProvider.get_map()
         self.timeout = timeout
 
-        self._direction = 'left'
+        self._direction = "left"
 
-        self._green_light_delay = 5  # Wait before the ego's lane traffic light turns green
+        self._green_light_delay = (
+            5  # Wait before the ego's lane traffic light turns green
+        )
         self._flow_tl_dict = {}
         self._init_tl_dict = {}
 
         self._end_distance = 10  # Distance after the junction before the scenario ends
 
-        self._flow_speed = get_value_parameter(config, 'flow_speed', float, 20)
-        self._source_dist_interval = get_interval_parameter(config, 'source_dist_interval', float, [25, 50])
+        self._flow_speed = get_value_parameter(config, "flow_speed", float, 20)
+        self._source_dist_interval = get_interval_parameter(
+            config, "source_dist_interval", float, [25, 50]
+        )
         self._scenario_timeout = 240
 
[…]
