### 2026-06-04 10c7b0e
refactored current error handling and added a switched to a more robust method of launching CARLA
