+++ cawsr.py
@@ -275,34 +275,41 @@ class AWScenarioRunner(object):
                 "Could not load scenario. Please check if the agent class is loading correctly."
             )
             result = False
-
-        self.carla_client.stop_recorder()
+        finally:
+            self.carla_client.stop_recorder()
 
         # stop the MetricsCollector thread
         MetricsCollector.reset()
 
-        # analyse the scenario
-        criteria = self._output_criteria(
-            self.scenario_manager.scenario.get_criteria(),  # type: ignore
-            f"{self.results_manager.last_scenario}/{scenario_name}.json",
-        )
-        logger.info("Calculating driving score...")
-        driving_score = self._calculate_driving_score(criteria)
+        try:
+            # analyse the scenario, throws exception if scenario didn't finish
+            criteria = self._output_criteria(
+                self.scenario_manager.scenario.get_criteria(),  # type: ignore
+                f"{self.results_manager.last_scenario}/{scenario_name}.json",
+            )
+            logger.info("Calculating driving score...")
 
-        result_dict = result_.get()
-        result_dict["status"] = result
-        result_dict["driving_score"] = driving_score
+            result_dict = result_.get()
+            result_dict["driving_score"] = self._calculate_driving_score(criteria)
+            result_dict["status"] = result
+
+        except Exception:
+            logger.info("Something went wrong, retrying scenario...")
 
         # read the scenario definition
         if algorithm_mode:
             algorithm._update_generator(seed)  # type: ignore
 
-            definition = algorithm._scenario_callback(  # type: ignore
-                current_definiton, driving_score
-            )
-            result_dict["definition"] = definition
-
-        result_.put(result_dict)
+            try:
+                definition = algorithm._scenario_callback(  # type: ignore
+                    current_definiton, result_dict["driving_score"]
+                )
+                result_dict["definition"] = definition
+                result_.put(result_dict)
+            except Exception:
+                logger.error(
+                    "Something went wrong while processing algorithm callback; is CARLA alive?"
+                )
 
     def _tick_carla(self) -> None:
         """Advances CARLA 1 tick into the future"""
@@ -372,7 +379,7 @@ class AWScenarioRunner(object):
                 0
             ]  # route id. Multiple routes currently aren't supported, so use first route
 
-            json_definition = self._cawsr_process(
+            result = self._cawsr_process(
                 route_config=route_config,
                 env_config=env_config,
                 scenario_name=scenario.stem,
@@ -383,7 +390,7 @@ class AWScenarioRunner(object):
             )
 
             self.results_manager.parse_json(
-                json_definition,  # type: ignore
+                result["definition"],  # type: ignore
                 scenario.stem,
                 str(i),
                 save_def=True,
@@ -396,42 +403,56 @@ class AWScenarioRunner(object):
 
         runs = len(scenarios)
         for i in range(0, runs):
-            logger.info(f"Running scenario {i + 1}/{runs}")
+            retry_attempts = 2
+            attempts = 0
 
-            if self._conf["benchmark"]["random_sampling"]:
-                scenario = random.choice(scenarios)
-                scenarios.remove(scenario)  # each scenario can only be executed once
-            else:
-                scenario = scenarios[i]
+            while attempts < retry_attempts:
+                logger.info(f"Running scenario {i + 1}/{runs}")
 
-            self._load_scenario(
-                scenario_name=scenario.stem,
-                path=str(scenario.absolute()),
-                run=i,
-                save_def=True,
-            )
+                if self._conf["benchmark"]["random_sampling"]:
+                    scenario = random.choice(scenarios)
+                    scenarios.remove(
+                        scenario
+                    )  # each scenario can only be executed once
+                else:
[…]
