### 2026-06-04 9a1546b
refactored current error handling and added a switched to a more robust method of launching CARLA
