### 2026-06-04 6533a9c
added additional log statements
