### 2026-06-04 1390d86
fixed bug with retry not being triggerd on CARLA crash
