+++ cawsr.py
@@ -18,6 +18,7 @@ import random
 import json
 
 import multiprocessing
+import multiprocessing.queues
 
 import carla
 
@@ -136,224 +137,221 @@ class AWScenarioRunner(object):
             current_definiton (Union[dict, None]): Optional dict containing the current scenario definition, used for algorithmic scenario generation
             algorithm (Union[Callable, None]): Optional Callable algorithm class, used for algorithmic scenario generation. See docs for expected interface.
         """
-        logger.info("Initialising Scenario Manager...")
-        self.scenario_manager = ScenarioManager(
-            self.DEBUG,
-            self._carla.SYNC,
-            self._carla.TIMEOUT,
-        )
-
-        logger.info("Starting the MetricsCollector thread...")
-
-        MetricsCollector.init_state(
-            metrics_collected,
-            os.path.join(self.results_manager.last_scenario, "execution_time.txt"),
-            include=False,
-        )
-
-        logger.info("Connecting to client...")
-        self.carla_client = carla.Client(self._carla.HOST, self._carla.PORT)
-        self.carla_client.set_timeout(self._carla.TIMEOUT)
-        CarlaDataProvider.set_client(self.carla_client)
-
-        logger.info("Fetching current world...")
-
-        self.carla_world = self.carla_client.get_world()
-        logger.info("Updating map...")
-        self.carla_client.load_world(env_config.town)
-
-        logger.info("Updating world settings:")
-
-        # tick asynchronously until then
-        settings = self.carla_world.get_settings()
-        settings.synchronous_mode = True
-        settings.fixed_delta_seconds = self._carla.FIXED_DELTA_SECONDS
-        self.carla_world.apply_settings(settings)
-
-        logger.info(f"{settings.__str__()}")
-
-        logger.info("Restarting GameTime...")
-        GameTime.restart()
-
-        # update the world
-        CarlaDataProvider.set_world(self.carla_world)
-
-        # attempt to spawn the ego 3 times for redundancy
-        max_attempts = 3
-        for attempt in range(max_attempts):
-            try:
-                logger.info("Spawning ego...")
-                ego = EgoVehicle(env_config)
-                actor = ego.spawn()
-                self.ego_vehicles.append(actor)
-                logger.info(f"Spawned ego with id: {actor.id}")
-                break  # exit loop if successful
-            except Exception as e:
-                logger.error(f"Failed to spawn ego on attempt {attempt + 1}: {e}")
-                if attempt == max_attempts - 1:
-                    logger.error(
-                        "Failed to spawn ego: Terminating scenario. This is most likely an issue with CARLA"
-                    )
-                    raise
-                time.sleep(2)
-
-        # client must tick to spawn actors
-        self._tick_carla()
-
-        logger.info("Initialising Autoware...")
-        agent_class_name = self.module_aw_agent.__name__.title().replace("_", "")
-        try:
-            logger.info(getattr(self.module_aw_agent, agent_class_name))
-            self.aw_agent = getattr(self.module_aw_agent, agent_class_name)(
-                env_config
-            )  # agent init function
-            route_config.agent = self.aw_agent
-        except Exception as e:  # Forces the simulation to run synchronously # pylint: disable=broad-except
-            logger.error("Could not setup required agent due to {}".format(e))
-            self._cleanup()
-            result = False
-            return
-
-        logger.info("Loading route...")
-
-        gps_route, route = route_manipulation.interpolate_trajectory(
-            route_config.keypoints
-        )
-        route_config.agent.set_global_plan(gps_route, route)  # set agent route
-
-        # visualise the route in CARLA and set location to first waypoint
-        visualise_route([waypoint[0] for waypoint in route])
[…]
