### 2026-06-04 7ebc59b
added check to only cleanup scenario if executed successfully
