### 2026-06-04 5acabe3
updated paper to match JOSS scope
