+++ paper.md
@@ -67,39 +67,16 @@ Therefore, `CAWSR` is designed to minimise such nondeterminism throughout the ev
 
 # State of the Field
 
-Simulation-based testing of Autonomous Vehicles spans several tools and frameworks, each targeting different aspects of the evaluation pipeline. CAWSR occupies a unique niche at the intersection of industry-grade ADS integration, scenario-based testing, and modern simulator support.
+Simulation-based testing of Autonomous Vehicles spans several tools, but CAWSR occupies a unique niche at the intersection of industry-grade ADS integration, scenario-based testing, and programmatic scenario generation.
 
-### Simulation Platforms
+**Simulators**: CARLA [@carla_sim] and SR [@carla_scenario_runner_2025] play a key role as the de-facto standard for open-source ADS research. However, their driving agent architecture inherently assumes a black-box model (for a good reason to encapsulate the ADS details), lacking the necessary mechanism to manage the complex massage passing of a ROS2-based system like Autoware.
 
-CARLA [@carla_sim] has established itself as the standard open-source simulator for ADS research, offering a rich ecosystem of tools, high-fidelity sensor models, and an active community. Its closest historical competitor for industry-grade ADS testing was LGSVL [@9294422], a Unity-based simulator developed by LG Electronics that provided integration with Apollo. However, LGSVL's official support was discontinued in January 2022, leaving a significant gap for researchers who relied on it as the simulation backend for their testing pipelines.
+**Autoware Integrations**: The CARLA-Autoware Bridge [@carlaautowarebridge] provides low-level data translation between CARLA and ROS2, serving as a vital dependency for our work. However, it does not provide any scenario execution capabilities. PCLA [@tehrani2025pcla] simplifies the deployment of multiple pretrained ML-based driving agents across multiple CARLA versions, but abstracts away the deep simulator-agent integration required for modular ADS like Autoware.
 
-### Scenario Execution Engines
+Recent concurrent developments, such as the autoware-carla leaderboard [@awcl_26], have also recognised this gap, introducing support for executing predefined Leaderboard scenarios with Autoware in CARLA. While these efforts confirm the urgent need, they optimise mainly for static benchmark execution. CAWSR diverges in its fundamental design by exposing a programmatic interface tailored for algorithmic, iterative scenario generation, which is a strict prerequisite for systematic ADS testing research.
 
-Within the CARLA ecosystem, Scenario Runner (SR) [@carla_scenario_runner_2025] is the standard scenario execution engine, underpinning the CARLA Leaderboard[@carla_leaderboard] evaluation framework. SR is well-suited to evaluating black-box ML-based agents that expose only sensor inputs and control outputs. However, it is not designed to interface with full-stack, industry-grade ADS such as Autoware, which rely on complex middleware architectures (e.g., ROS2) for internal communication. CAWSR extends the SR execution model specifically to accommodate Autoware's architecture, enabling scenario-based testing and evaluation of module-based ADS, which SR does not support.
+**Build vs. Contribute**: The state of the field presents a clear rationale for CAWSR as a new tool rather than a contribution to an existing project. Extending SR directly to support Autoware would require fundamentally rewriting its core agent model, breaking backwards compatibility for ML-based users. Therefore, building CAWSR as a standalone orchestration layer that leverages SR's underlying behavior trees, while maintaining a specialised ROS2 agent interface, presents the most viable scholarly contribution. 
 
-### Autoware Integration Tools
-
[…]
